fix imports; shorten text from api to base
