+++ src/apps/comments/models.py
@@ -1,6 +1,6 @@
-from django.db import models
 from django.conf import settings
 from django.core.validators import MaxLengthValidator
+from django.db import models
 
 
 class Comment(models.Model):
