+++ src/base/services.py
@@ -1,4 +1,3 @@
-from django.contrib.sessions.backends import file
 from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
 
@@ -15,3 +14,13 @@ def user_avatar_path(instance, filename):
     # id будет указываться как None
 
     return f"avatar/user_{instance.id}/{filename}"
+
+
+def shorten_text(full_text, n) -> str:
+    """
+    Shorten text to n characters with rounding by last word
+    """
+    short_text = full_text[:100]
+    if len(full_text) > n and full_text[n] != "":
+        short_text = short_text[: short_text.rfind(" ")]
+    return short_text
