+++ src/tests/conftest.py
@@ -1,10 +1,10 @@
 import pytest
-from src.apps.recipes.models import Recipe
 from django.core.management import call_command
-from rest_framework.test import APIClient
-from src.apps.recipes.models import Recipe
-from src.apps.ingredients.models import Ingredient, Unit
 from django.utils.text import slugify
+from rest_framework.test import APIClient
+
+from src.apps.ingredients.models import Ingredient, Unit
+from src.apps.recipes.models import Recipe
 
 
 @pytest.fixture(scope="session")
