+++ src/tests/ingredients/test_ingredients.py
@@ -1,4 +1,5 @@
 import pytest
+
 from src.apps.ingredients.models import Unit, Ingredient, IngredientInRecipe
